Separate CharRegexMatchTests
